+++ solutions/google-code-jam/2017/qualification-round/oversized-pancake-flipper/oversized-pancake-flipper.cpp
@@ -1,12 +1,12 @@
 #include <iostream>
 
 int main() {
-    int t, flipper_capacity;
+    int16_t t, flipper_capacity;
     std::string pancakes;
 
     std::cin >> t;
 
-    for (int i = 0; i < t; i++) {
+    for (int16_t i = 0; i < t; i++) {
         std::cin >> pancakes >> flipper_capacity;
     }
