+++ solutions/project-euler/009/009.cpp
@@ -1,11 +1,11 @@
 #include <cmath>
 #include <iostream>
 
-int s = 1000;
-int s2 = s / 2;
-int limit = sqrt(s2) - 1;
+int16_t s = 1000;
+int16_t s2 = s / 2;
+int16_t limit = sqrt(s2) - 1;
 
-int gcd(int x, int y) {
+int16_t gcd(int16_t x, int16_t y) {
     if (y == 0) {
         return x;
     }
@@ -13,7 +13,7 @@ int gcd(int x, int y) {
 }
 
 int main() {
-    int m, n, sm, k, d, a, b, c;
+    int16_t m, n, sm, k, d, a, b, c;
 
     for (m = 2; m <= limit; m++) {
         if (s2 % m == 0) {
