+++ solutions/project-euler/010/010.cpp
@@ -1,7 +1,8 @@
 #include <cmath>
+#include <cstdint>
 #include <iostream>
 
-bool is_prime(unsigned long n) {
+bool is_prime(uint32_t n) {
     if (n == 2) {
         return true;
     }
@@ -9,7 +10,7 @@ bool is_prime(unsigned long n) {
         return false;
     }
 
-    for (unsigned long i = 3; i <= sqrt(n); i += 2) {
+    for (uint32_t i = 3; i <= sqrt(n); i += 2) {
         if (n % i == 0) {
             return false;
         }
@@ -18,8 +19,8 @@ bool is_prime(unsigned long n) {
 }
 
 int main() {
-    unsigned int i;
-    unsigned long long sum = 2;
+    uint32_t i;
+    uint64_t sum = 2;
 
     for (i = 3; i < 2000000; i += 2) {
         if (is_prime(i)) {
