+++ solutions/uri/1008/1008.cpp
@@ -1,7 +1,8 @@
+#include <cstdint>
 #include <cstdio>
 
 int main() {
-    int a, b;
+    int16_t a, b;
     float c;
 
     scanf("%d", &a);
