+++ solutions/uri/1010/1010.cpp
@@ -2,7 +2,7 @@
 #include <cstdio>
 
 int main() {
-    int a, b;
+    int16_t a, b;
     float c, s = 0.0;
 
     while (scanf("%d %d %f", &a, &b, &c) != EOF) {
