+++ solutions/uri/1042/1042.cpp
@@ -1,7 +1,8 @@
+#include <cstdint>
 #include <cstdio>
 
 int main() {
-    int a, b, c, g, m, s, x;
+    int16_t a, b, c, g, m, s, x;
 
     scanf("%d %d %d", &a, &b, &c);
