+++ solutions/uri/1072/1072.cpp
@@ -1,9 +1,12 @@
+#include <cstdint>
 #include <cstdio>
 #include <iostream>
 
 int main() {
-    int c, n, i = 0, o = 0;
+    int16_t c, n, i = 0, o = 0;
+
     std::cin >> c;
+
     while (c--) {
         std::cin >> n;
         if (n > 9 && n < 21) {
@@ -15,5 +18,6 @@ int main() {
 
     std::cout << i << " in" << std::endl;
     std::cout << o << " out" << std::endl;
+
     return 0;
 }
