+++ solutions/uri/1075/1075.cpp
@@ -1,13 +1,17 @@
+#include <cstdint>
 #include <cstdio>
 #include <iostream>
 
 int main() {
-    int n;
+    int16_t n;
+
     std::cin >> n;
-    for (int i = 1; i <= 10000; i++) {
+
+    for (int16_t i = 1; i <= 10000; i++) {
         if (i % n == 2) {
             printf("%d\n", i);
         }
     }
+
     return 0;
 }
